requirements.txt
asgiref==3.7.2
Django==5.0.1
django-phonenumber-field==7.3.0
django-rest-registration==0.8.3
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
phonenumbers==8.13.28
pillow==10.2.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
